moved byte parsing functions to cc file
