+++ tensorflow_io/core/BUILD
@@ -186,6 +186,7 @@ cc_library(
         "kernels/bigtable/bigtable_row_set.h",
         "kernels/bigtable/bigtable_version_filters.cc",
         "kernels/bigtable/bigtable_version_filters.h",
+        "kernels/bigtable/serialization.cc",
         "kernels/bigtable/serialization.h",
         "ops/bigtable_ops.cc",
     ],
